Completes bits.rb, comments out classless code.
Adds Bits#to_a and Bits#sum methods, fixes an error with accidentally
calling the instance version of power_of_two? rather than newer class
method version.

Bits.pattern returns the regex of a bitfield as it did before, but
employs Bits.insert, which returns the regex of a bitfield without
the start/end anchors so it can be interpolated into other regexes.