Changes varea.rb to use new Bits.rb class
bparse(str) was replaced with Bits.new(str), and instead of
bparse returning nil, Bits#error? was check to see if the bits were
all powers of 2.

The bitfield regexes were all changed to call Bits.pattern or
Bits.insert, depending on whether the pattern was on its own or
inserted inside a larger regex.