Doesn't snip off first line of section at first
Varea doesn't remove the first line of a section body before sticking it
into the contents of a section class. That is now the responsibility of
each class. (It allows for the current_line variable to be incrememted
prior to parsing a line rather than after, which means fewer times
needed to write the expression to increment it)